updata test_qeq

+++ dmff/settings.py
@@ -1,7 +1,7 @@
 from jax.config import config
 
-PRECISION = 'double'  # 'double'
-#PRECISION = 'single'  # 'double'
+#PRECISION = 'double'  # 'double'
+PRECISION = 'single'  # 'double'
 
 DO_JIT = True
